app/whisper-api/requirements.txt
flask
openai
python-dotenv
flask_cors
mangum
click
pydub
SpeechRecognition